% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{create_table}
\alias{create_table}
\title{Converts a df into a table grob and prints turns it into a png placed in the destination}
\usage{
create_table(df, title, destination, name, dimensions)
}
\arguments{
\item{df}{A dataframe to be converted into a table}

\item{title}{A character string representing the title to place above the table}

\item{destination}{A character string representing the filepath to place the png}

\item{name}{A character string representing the name of the png}

\item{dimensions}{A length 2 list containing the width and height of the png represented as numbers (width first)}
}
\value{
Nothing
}
\description{
Converts a df into a table grob and prints turns it into a png placed in the destination
}
\examples{
\dontrun{
create_table(
  data.frame(
  Age = c(25, 30, 35),
  Score = c(85.5, 90.0, 78.3)
  ),
  "TableTitle",
  "tests/testdir",
  "MyTableImage",
  list(1500,1000)
)
}
}
