% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{generate_scatterplots}
\alias{generate_scatterplots}
\title{Create scatterplots to compare all variables against a target variable}
\usage{
generate_scatterplots(vars, plotdata, target, numberfrom)
}
\arguments{
\item{vars}{A character vector of the names of the numeric variables inside data}

\item{plotdata}{A tibble or data frame that contains the values for all of the variables}

\item{target}{A character string representing the variable of interest that we will plot all the other variables against}

\item{numberfrom}{A number that represents the starting figure number of the first scatterplot}
}
\value{
A list of scatterplots
}
\description{
Create scatterplots to compare all variables against a target variable
}
\examples{
\dontrun{
generate_scatterplots(
  c("Age","Score"),
  data.frame(
  Age = c(25, 30, 35),
  Score = c(85.5, 90.0, 78.3)
  ),
  "Age",
  5
)
}
}
