% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{save_cleaned_data}
\alias{save_cleaned_data}
\title{Save cleaned data to a CSV file}
\usage{
save_cleaned_data(data, destination)
}
\arguments{
\item{data}{A tibble containing the cleaned data}

\item{destination}{A character string containing the file path where the data should be saved}
}
\value{
NULL (invisible)
}
\description{
Save cleaned data to a CSV file
}
\examples{
\dontrun{
save_cleaned_data(cleaned_data, "data/processed/cleaned_data.csv")
}
}
